--- a/src/storage/mod.rs
+++ b/src/storage/mod.rs
@@ -2,10 +2,12 @@ mod archive_index;
 mod compression;
 mod database;
 mod s3;
+mod sqlite_pool;
 
 pub use self::compression::{compress, decompress, CompressionAlgorithm, CompressionAlgorithms};
 use self::database::DatabaseBackend;
 use self::s3::S3Backend;
+use self::sqlite_pool::SqliteConnectionPool;
 use crate::error::Result;
 use crate::web::metrics::RenderingTimesRecorder;
 use crate::{db::Pool, Config, Metrics};
@@ -13,6 +15,7 @@ use anyhow::{anyhow, ensure};
 use chrono::{DateTime, Utc};
 use path_slash::PathExt;
 use std::io::BufReader;
+use std::num::NonZeroU64;
 use std::{
     collections::{HashMap, HashSet},
     ffi::OsStr,
@@ -113,6 +116,7 @@ enum StorageBackend {
 pub struct Storage {
     backend: StorageBackend,
     config: Arc<Config>,
+    sqlite_pool: SqliteConnectionPool,
 }
 
 impl Storage {
@@ -123,6 +127,10 @@ impl Storage {
         runtime: Arc<Runtime>,
     ) -> Result<Self> {
         Ok(Storage {
+            sqlite_pool: SqliteConnectionPool::new(
+                NonZeroU64::new(config.max_sqlite_pool_size)
+                    .ok_or_else(|| anyhow!("invalid sqlite pool size"))?,
+            ),
             config: config.clone(),
             backend: match config.storage_backend {
                 StorageKind::Database => {
@@ -229,7 +237,9 @@ impl Storage {
 
     pub(crate) fn exists_in_archive(&self, archive_path: &str, path: &str) -> Result<bool> {
         match self.get_index_filename(archive_path) {
-            Ok(index_filename) => Ok(archive_index::find_in_file(index_filename, path)?.is_some()),
+            Ok(index_filename) => {
+                Ok(archive_index::find_in_file(index_filename, path, &self.sqlite_pool)?.is_some())
+            }
             Err(err) => {
                 if err.downcast_ref::<PathNotFoundError>().is_some() {
                     Ok(false)
@@ -306,8 +316,12 @@ impl Storage {
         if let Some(ref mut t) = fetch_time {
             t.step("find path in index");
         }
-        let info = archive_index::find_in_file(self.get_index_filename(archive_path)?, path)?
-            .ok_or(PathNotFoundError)?;
+        let info = archive_index::find_in_file(
+            self.get_index_filename(archive_path)?,
+            path,
+            &self.sqlite_pool,
+        )?
+        .ok_or(PathNotFoundError)?;
 
         if let Some(t) = fetch_time {
             t.step("range request");