Coming from the performance issues with the archive index I took some time to update the whole topic. 

While I started to dig into optimizing the CBOR reader I then thought about the fact that CBOR will never be scalable for bigger crates, so I started looking for alternatives  that would include some indexing. 

I did some high level tests with LMDB, sled and rocksdb, but settled for a boring, old technology: SQLite. 

( yes, I'm definitely biased towards older but perhaps more reliable things). 

In my mind, this would give us a reliable, fast, single-file format, which would also be easily extensible if we need to add more information. File-size is slightly higher, but IMO not an issue. 

This PR would only use the new format for new builds, while I would plan for a slow (lazy or active) migration of the other archive-stored releases. 


### micro-benchmarks: 
using the archive index for `stm32ral` (~1.2 million files)

```

with open               time:   [59.434 µs 59.604 µs 59.810 µs]
                        change: [-1.5928% -0.5698% +0.3278%] (p = 0.27 > 0.05)
                        No change in performance detected.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe

without open            time:   [7.0644 µs 7.1273 µs 7.2172 µs]
                        change: [+2.2772% +3.9894% +5.9668%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 16 outliers among 100 measurements (16.00%)
  7 (7.00%) high mild
  9 (9.00%) high severe
```

- "with open" -> including creating a new SQLite connection opening the file 
- "without open" -> reusing an existing connection. 
- for comparison: reading an old cbor-index of that size takes ~20 seconds 


That speed doesn't fully scale with parallelism, but IMO should be good enough, especially compared to the current speed. 

## deployment / next steps 
This PR intentionally only uses the new format for new builds, so i anything goes wrong we'll be able to revert & rebuild. 

A next step could be to write an _index repack_ command that would convert old existing indexes so we can remove the old cbor code & benefit from the performance improvements.