@syphar 2023-03-03
I would love some feedback on this improvement, 

cc @rust-lang/docs-rs