@Nemo157 2023-03-03
The other interesting metric would be size, both before and after compression, I assume it'll be a little larger to store the index.

Another option discussed ages ago was [`fst`](https://docs.rs/fst/latest/fst/) which essentially is just the index part without any data storage (it provides an API sort of like `Map<Path, u64>`). We could combine that with something like a series of individually CBOR encoded `FileInfo`s and have `fst` to lookup the offset on where the `FileInfo` starts.