@syphar 2023-03-03
I'll answer in more detail later, one thought already was that sqlite is much easier to debug since you have db and cli tools that can easily edit / change the database.