@syphar 2023-03-07
what do you believe is missing? I intentionally would only do the new format for new releases, so we can rollback and rebuild. 

Next step if that works would be adding some migration for old releases, so we are safer against the crawler / high load.