@jsha 2023-03-07
Overall I'm in favor! 👍🏻

I had to rehydrate some mental state on this, so writing it down for others who might find it useful:

The archive index was first implemented as part of https://github.com/rust-lang/docs.rs/pull/1342. My first question was "doesn't the zip central directory solve this?" And in fact that's discussed in that PR:

> in the beginning I was actually thinking about using the zip central directory as the index, but that lead to some complexity (while keeping the file hashes). Here in this project the custom index, integrated with the existing compression logic made more sense.

> The ZIP directory is just the end of the file. It [contains all the filenames, and also the positions to fetch](https://en.wikipedia.org/wiki/ZIP_(file_format)#Central_directory_file_header). I had a working prototype that had a kind of virtual buffer which prefetched the end of the archive and then the range for the file. After that worked, it seemed more complex than just using our own index, and decompressing the stream ourselves.

Also, doing a little more research, one problem with the zip central directory is that it occurs at the end of the file and its beginning position is not known - you have to seek backwards from the end of the file to find it.

Also, if the problem we're dealing with here is that the index itself is so large that decompressing / parsing the whole index is too expensive, then the zip central directory has the same problem - we would have to parse the whole thing to find the file we want. It would be possible to generate a zip central directory in sorted order and binary search across it, but because directory entries are not of constant size this would get very messy very quickly.

Similarly, we could make the CBOR index sorted and do binary search, but it presumably has the same problem - variable length entries to to path name. Though we could maybe solve that by storing filename hashes instead of filenames? Still, I see why it makes sense to switch to SQLite, which has indexing of data by strings as a first class feature, rather than reinventing indexes files for ourselves from scratch. 👍🏻 Another longstanding and simple technology that fills the same niche is SSTables, which has a Rust implementation: https://crates.io/crates/sstable.

For thoroughness, another approach we could take is to turn the problem inside-out. Right now we generate zip files as our first-class storage, and offer them for direct download (cheap for us) and serve webpages from them (expensive for us, particularly for large crates).

But fetching webpages is a frequent thing, and the world expects it to be cheap (thus, crawlers). Downloading zip files is a less frequent thing, and while it's good for it to be cheap, it's not so bad if it's a little expensive.

So what if we changed our archive storage format directly to something like SQLite or SSTable that has good performance and straightforward implementation when fetching up single files? Then when serving a request for a zip file we would convert the SQLite or SSTable to zip on the file, possibly with a caching layer.