@jsha 2023-03-07
Also, what's the distribution of the next-biggest crates after stm32ral? Is it reasonable to use /robots.txt to tell people not to bother crawling stm32ral and some other handful of mega-crates? It seems like we could spend a lot of engineering trying to make it fast to serve single pages from mega-crates in various ways.