@syphar 2023-03-08
> Also, what's the distribution of the next-biggest crates after stm32ral? Is it reasonable to use /robots.txt to tell people not to bother crawling stm32ral and some other handful of mega-crates? It seems like we could spend a lot of engineering trying to make it fast to serve single pages from mega-crates in various ways.

I once did this analysis based on a (~2021) inventory of our S3 bucket, when trying to debug bigger index performance :) it came down to ~200 crates with more than an 100k files in their docs, but goes down to only ~10 crates with more than 500k files. 



> For thoroughness, another approach we could take is to turn the problem inside-out. Right now we generate zip files as our first-class storage, and offer them for direct download (cheap for us) and serve webpages from them (expensive for us, particularly for large crates).
> 
> But fetching webpages is a frequent thing, and the world expects it to be cheap (thus, crawlers). Downloading zip files is a less frequent thing, and while it's good for it to be cheap, it's not so bad if it's a little expensive.
> 
> So what if we changed our archive storage format directly to something like SQLite or SSTable that has good performance and straightforward implementation when fetching up single files? Then when serving a request for a zip file we would convert the SQLite or SSTable to zip on the file, possibly with a caching layer.

This is definitely a valid idea. 

The main advantage of using the archive as we have it is that we can to S3 range requests to fetch single files out of the archive. I assume with SQLite for the docs, or SSTables, we would have to fetch the whole doc-package on the first request, especially when we account for the new ECS setup where we will scale up / down all the time. 
The bigger doc-archives can reach multiple gigabytes.

In all the design discussions about archive-storage , downloadable docs where actually only second priority, more a nice-to-have. 


> Overall I'm in favor! 👍🏻 

thank you for taking the time for the thorough comment & research!