@jsha 2023-03-08
>  I assume with SQLite for the docs, or SSTables, we would have to fetch the whole doc-package on the first request,

SSTables have an index, which should make it possible to fetch just the index and use range requests to get a specific file. But looking again at the sstable crate, it was last released over a year ago, so maybe not a great choice.

It's a bit of a bummer that the current approach requires reading the whole SQLite _index_ before making other fetches, since 291MB is still a lot of data to process. Speaking of which, your benchmark shows that fetching an entry from the 291MB index (including decompression?) takes about 60 _micro_seconds? That seems impossible.